support general data

+++ src/main/webapp/GetSummary.jsp
@@ -27,20 +27,27 @@
                     dataDir = " C:\\Software\\processed\\";
                     jarLocation = " C:\\Software\\CMIDemo\\src\\main\\webapp\\WEB-INF\\lib\\XdrHttp-1.0-SNAPSHOT.jar ";
                }
-              String execStr = java + " -cp " + jarLocation + " Demo "
-               + dataDir + " " + request.getParameter("lat") + " " + request.getParameter("lng") + " "
-               + request.getParameter("radius") + " " +
-               request.getParameter("absolute_hour") + " " + request.getParameter("mcc");
-              Process p = Runtime.getRuntime().exec(execStr);
-              p.waitFor();
-              BufferedReader bri = new BufferedReader
-                              (new InputStreamReader(p.getInputStream()));
-              while ((line = bri.readLine()) != null) {
-                if (line.trim().length() > 0) { // skip empty lines. not sure whey they are there.
-                    out.println(line);
-                }
+              String input = String.format(dataDir + request.getParameter("date") + "%02d" + ".csv",
+                Integer.parseInt(request.getParameter("hour")));
+
+              if (new File(input).exists()) {
+                  String execStr = java + " -cp " + jarLocation + " Demo "
+                   + input + " " + request.getParameter("lat") + " " + request.getParameter("lng") + " "
+                   + request.getParameter("radius") + " " + request.getParameter("mcc");
+                  // out.println(execStr);
+                  Process p = Runtime.getRuntime().exec(execStr);
+                  p.waitFor();
+                  BufferedReader bri = new BufferedReader
+                                  (new InputStreamReader(p.getInputStream()));
+                  while ((line = bri.readLine()) != null) {
+                    if (line.trim().length() > 0) { // skip empty lines. not sure whey they are there.
+                        out.println(line);
+                    }
+                  }
+                 bri.close();
+              } else {
+                out.println("Error: Data File Does Not Exist");
               }
-              bri.close();
             }
             catch (Exception err) {
               err.printStackTrace();
